+++ CHANGES.rst
@@ -2,6 +2,13 @@
  Changes for mkvenv
 ====================
 
+0.1.2
+=====
+
+ * installs to system python if --venv not specified
+ * install performs --upgrade
+ * --version writes to stdout
+
 0.1.1
 =====